clean up and divide in multiple components
